+++ packages/components/src/components/SegmentedControl/SegmentedControl.tsx
@@ -11,10 +11,10 @@ import { getContainerBreakpointSizeClassName } from "@/lib/getContainerBreakpoin
 import { type PropsContext } from "@/lib/propsContext";
 import { PropsContextProvider } from "@/lib/propsContext";
 import { TunnelExit, TunnelProvider } from "@mittwald/react-tunnel";
-import { FieldError } from "@/components/FieldError";
 import clsx from "clsx";
 import { useObjectRef } from "@react-aria/utils";
 import { useMakeFocusable } from "@/lib/hooks/dom/useMakeFocusable";
+import { useFieldComponent } from "@/lib/hooks/useFieldComponent";
 
 export interface SegmentedControlProps
   extends PropsWithChildren<Omit<Aria.RadioGroupProps, "children">>,
@@ -31,6 +31,13 @@ export const SegmentedControl = flowComponent("SegmentedControl", (props) => {
     ...rest
   } = props;
 
+  const {
+    FieldErrorView,
+    fieldPropsContext,
+    fieldProps,
+    FieldErrorResetContext,
+  } = useFieldComponent(props);
+
   const rootClassName = clsx(
     formFieldStyles.formField,
     styles.segmentedControlContainer,
@@ -39,43 +46,35 @@ export const SegmentedControl = flowComponent("SegmentedControl", (props) => {
   );
 
   const propsContext: PropsContext = {
-    Label: {
-      className: formFieldStyles.label,
-    },
-    FieldDescription: {
-      className: formFieldStyles.fieldDescription,
-      tunnelId: "fieldDescription",
-    },
-    FieldError: {
-      className: formFieldStyles.customFieldError,
-      tunnelId: "fieldError",
-    },
     Segment: {
       tunnelId: "segments",
       className: styles.segment,
     },
+    ...fieldPropsContext,
   };
 
   const localRadioRef = useObjectRef(ref);
   useMakeFocusable(localRadioRef);
 
   return (
-    <Aria.RadioGroup {...rest} className={rootClassName} ref={localRadioRef}>
-      <PropsContextProvider dependencies={["segment"]} props={propsContext}>
-        <TunnelProvider>
-          {children}
-
-          <div className={styles.segmentedControl}>
-            <div className={styles.segments}>
-              <TunnelExit id="segments" />
+    <Aria.RadioGroup
+      {...rest}
+      className={clsx(rootClassName, fieldProps.className)}
+      ref={localRadioRef}
+    >
+      <TunnelProvider>
+        <FieldErrorResetContext>
+          <PropsContextProvider dependencies={["segment"]} props={propsContext}>
+            <div className={styles.segmentedControl}>
+              <div className={styles.segments}>
+                <TunnelExit id="segments" />
+              </div>
             </div>
-          </div>
-
-          <TunnelExit id="fieldDescription" />
-          <TunnelExit id="fieldError" />
-        </TunnelProvider>
-      </PropsContextProvider>
-      <FieldError className={formFieldStyles.fieldError} />
+            {children}
+          </PropsContextProvider>
+        </FieldErrorResetContext>
+        <FieldErrorView />
+      </TunnelProvider>
     </Aria.RadioGroup>
   );
 });