+++ packages/components/src/integrations/react-hook-form/components/Field/stories/NumberField.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import React from "react";
+import React, { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import { action } from "storybook/actions";
 import { Label } from "@/components/Label";
@@ -10,6 +10,7 @@ import { ActionGroup } from "@/components/ActionGroup";
 import { sleep } from "@/lib/promises/sleep";
 import { NumberField } from "@/components/NumberField";
 import { FieldDescription } from "@/components/FieldDescription";
+import { FieldError } from "@/components/FieldError";
 
 const submitAction = action("submit");
 
@@ -88,13 +89,46 @@ type Story = StoryObj<typeof Field>;
 
 export const Default: Story = {};
 
-export const WithFocusAndError: Story = {
+export const WithFieldError: Story = {
   render: () => {
-    const form = useForm();
+    const form = useForm({
+      defaultValues: {
+        field: "",
+      },
+    });
+    useEffect(() => {
+      form.setError("field", {
+        type: "required",
+        message: "ErrorFromForm",
+      });
+    }, []);
 
     return (
       <Form form={form} onSubmit={async () => await sleep(2000)}>
-        <Field name={"text"} rules={{ required: true }}>
+        <Field name="field">
+          <NumberField>
+            <Label>Age</Label>
+          </NumberField>
+        </Field>
+        <NumberField isInvalid>
+          <Label>Age</Label>
+          <FieldError>ErrorFromOuterFieldError!</FieldError>
+        </NumberField>
+      </Form>
+    );
+  },
+};
+
+export const WithFocus: Story = {
+  render: () => {
+    const form = useForm({
+      defaultValues: {
+        field: "",
+      },
+    });
+    return (
+      <Form form={form} onSubmit={async () => await sleep(2000)}>
+        <Field name={"field"}>
           <NumberField>
             <Label>Age</Label>
           </NumberField>
@@ -103,7 +137,7 @@ export const WithFocusAndError: Story = {
         <Button
           onPress={() =>
             form.setError(
-              "text",
+              "field",
               { type: "required", message: "oh no" },
               { shouldFocus: true },
             )
@@ -111,7 +145,7 @@ export const WithFocusAndError: Story = {
         >
           err through form
         </Button>
-        <Button onPress={() => form.setFocus("text")}>
+        <Button onPress={() => form.setFocus("field")}>
           focus through form
         </Button>
         <Button type="submit">Submit</Button>